+++ test/generators/install_generator_test.rb
@@ -44,29 +44,29 @@ class InstallGeneratorTest < Rails::Generators::TestCase
   test "app" do
     run_generator
 
-    assert_file "app/lib/rodauth_app.rb", /class RodauthApp < Rodauth::Rails::App/
-    assert_file "app/lib/rodauth_app.rb", /configure do/
-    assert_file "app/lib/rodauth_app.rb", /:login, :logout, :remember,$/
-    assert_file "app/lib/rodauth_app.rb", /hmac_secret "[a-z0-9]{128}"/
-    assert_file "app/lib/rodauth_app.rb", /rails_controller { RodauthController }/
-    assert_file "app/lib/rodauth_app.rb", /flash_notice_key/
-    assert_file "app/lib/rodauth_app.rb", /Remember Feature/
-    assert_file "app/lib/rodauth_app.rb", /logout_redirect/
-    assert_file "app/lib/rodauth_app.rb", /rodauth\.load_memory/
+    assert_file "app/misc/rodauth_app.rb", /class RodauthApp < Rodauth::Rails::App/
+    assert_file "app/misc/rodauth_app.rb", /configure do/
+    assert_file "app/misc/rodauth_app.rb", /:login, :logout, :remember,$/
+    assert_file "app/misc/rodauth_app.rb", /hmac_secret "[a-z0-9]{128}"/
+    assert_file "app/misc/rodauth_app.rb", /rails_controller { RodauthController }/
+    assert_file "app/misc/rodauth_app.rb", /flash_notice_key/
+    assert_file "app/misc/rodauth_app.rb", /Remember Feature/
+    assert_file "app/misc/rodauth_app.rb", /logout_redirect/
+    assert_file "app/misc/rodauth_app.rb", /rodauth\.load_memory/
   end
 
   test "app with --json option" do
     run_generator %w[--json]
 
-    assert_file "app/lib/rodauth_app.rb", /:login, :logout, :remember, :json,$/
-    assert_file "app/lib/rodauth_app.rb", /only_json\? true/
+    assert_file "app/misc/rodauth_app.rb", /:login, :logout, :remember, :json,$/
+    assert_file "app/misc/rodauth_app.rb", /only_json\? true/
   end
 
   test "app with --jwt option" do
     run_generator %w[--jwt]
 
-    assert_file "app/lib/rodauth_app.rb", /:login, :logout, :jwt,$/
-    assert_file "app/lib/rodauth_app.rb", /jwt_secret "[a-z0-9]{128}"/
+    assert_file "app/misc/rodauth_app.rb", /:login, :logout, :jwt,$/
+    assert_file "app/misc/rodauth_app.rb", /jwt_secret "[a-z0-9]{128}"/
   end
 
   test "controller" do