+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## HEAD
 
+* Add `Rodauth::Rails::Auth.instance` which returns internal request's Rodauth instance (@janko)
+
+* Deprecate `Rodauth::Rails.rodauth` in favour of `Rodauth::Rails::Auth.instance` (@janko)
+
 * Generate named auth classes in `rodauth:install` generator (@janko)
 
 * Generate `rodauth_app.rb` in `app/misc` directory (@janko)